+++ seed/MyScripts/Jira/Jira Cache Assets.ps1
@@ -4,7 +4,7 @@
 # VariableSwitch: $Force, false, Wymuś odświeżenie|
 # SirkWorkflow: JiraAssetsCache
 # SirkSystemCredential: Jira
-# SirkJiraAssetAql: objectType in objectTypeAndChildren("Sprzęt użytkownika") OR objectType = "Users" OR object HAVING outboundReferences(objectType = "Users") OR object HAVING inboundReferences(objectType = "Users")
+# SirkJiraAssetAql: objectType in objectTypeAndChildren("Sprzęt użytkownika")
 # SirkJiraAssetLabelAttribute: Nazwa_sieciowa
 # MultiHost: false
 # runAsUser: 0